Modulo de patrimonio 50% comcluido

+++ database/migrations/tenant/2024_07_17_051007_create_audits_table.php
@@ -13,33 +13,36 @@ class CreateAuditsTable extends Migration
      */
     public function up()
     {
-
         $connection = config('audit.drivers.database.connection', config('database.default'));
         $table = config('audit.drivers.database.table', 'audits');
 
-        Schema::connection($connection)->create($table, function (Blueprint $table) {
-            $morphPrefix = config('audit.user.morph_prefix', 'user');
+        // Verifica se a tabela já existe antes de criá-la
+        if (!Schema::connection($connection)->hasTable($table)) {
+            Schema::connection($connection)->create($table, function (Blueprint $table) {
+                $morphPrefix = config('audit.user.morph_prefix', 'user');
 
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('company_id')->nullable(); // Campo para a empresa
-            $table->string($morphPrefix . '_type')->nullable(); // Tipo polimórfico
-            $table->unsignedBigInteger($morphPrefix . '_id')->nullable(); // ID polimórfico
-            $table->string('event');
-            $table->morphs('auditable');
-            $table->text('old_values')->nullable();
-            $table->text('new_values')->nullable();
-            $table->text('url')->nullable();
-            $table->ipAddress('ip_address')->nullable();
-            $table->string('user_agent', 1023)->nullable();
-            $table->string('tags')->nullable();
-            $table->timestamps();
+                $table->bigIncrements('id');
+                $table->unsignedBigInteger('company_id')->nullable(); // Campo para a empresa
+                $table->string($morphPrefix . '_type')->nullable(); // Tipo polimórfico
+                $table->unsignedBigInteger($morphPrefix . '_id')->nullable(); // ID polimórfico
+                $table->string('event');
+                $table->morphs('auditable');
+                $table->text('old_values')->nullable();
+                $table->text('new_values')->nullable();
+                $table->text('url')->nullable();
+                $table->ipAddress('ip_address')->nullable();
+                $table->string('user_agent', 1023)->nullable();
+                $table->string('tags')->nullable();
+                $table->timestamps();
 
-            $table->index([$morphPrefix . '_id', $morphPrefix . '_type']);
+                $table->index([$morphPrefix . '_id', $morphPrefix . '_type']);
 
-            // Adiciona chaves estrangeiras (opcional)
-            $table->foreign('company_id')->references('id')->on('companies')->onDelete('set null');
-        });
+                // Adiciona chaves estrangeiras (opcional)
+                $table->foreign('company_id')->references('id')->on('companies')->onDelete('set null');
+            });
+        }
     }
+
 
     /**
      * Reverse the migrations.
